fix(filter): cast GlobalAssetClass to int to guarantee integer serialization
System.Text.Json serializes boxed enum values in Dictionary<string, object>
as their string name when the runtime type is the enum itself. Explicitly
casting to int ensures the API always returns a numeric value that the
frontend filter can compare against.

+++ Financial.Application/Services/NavigationMapper.cs
@@ -66,7 +66,7 @@ internal static class NavigationMapper
                 ["ISIN"] = asset.ISIN,
                 ["Country"] = asset.Country,
                 ["LocalTypeCode"] = asset.LocalTypeCode,
-                ["GlobalAssetClass"] = asset.Class,
+                ["GlobalAssetClass"] = (int)asset.Class,
                 ["Quantity"] = asset.Quantity,
                 ["AveragePrice"] = asset.AveragePrice,
                 ["IsActive"] = asset.IsActive,
